+++ tests/framework/db/pgsql/conditions/InConditionBuilderTest.php
@@ -1,0 +1,68 @@
+<?php
+
+/**
+ * @link https://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license https://www.yiiframework.com/license/
+ */
+
+declare(strict_types=1);
+
+namespace yiiunit\framework\db\pgsql\conditions;
+
+use PHPUnit\Framework\Attributes\DataProviderExternal;
+use PHPUnit\Framework\Attributes\Group;
+use yii\base\InvalidArgumentException;
+use yii\db\conditions\InCondition;
+use yii\db\Query;
+use yiiunit\base\db\BaseDatabase;
+use yiiunit\framework\db\pgsql\conditions\providers\InConditionBuilderProvider;
+
+/**
+ * Unit test for {@see \yii\db\conditions\InConditionBuilder} with PostgreSQL driver.
+ *
+ * {@see InConditionBuilderProvider} for test case data providers.
+ *
+ * @author Wilmer Arambula <terabytesoftw@gmail.com>
+ * @since 2.2
+ */
+#[Group('db')]
+#[Group('condition')]
+#[Group('pgsql')]
+final class InConditionBuilderTest extends BaseDatabase
+{
+    protected $driverName = 'pgsql';
+
+    #[DataProviderExternal(InConditionBuilderProvider::class, 'buildCondition')]
+    public function testBuildCondition(array|object $condition, string $expected, array $expectedParams): void
+    {
+        $query = (new Query())->where($condition);
+
+        $db = $this->getConnection(true, false);
+
+        [$sql, $params] = $db->getQueryBuilder()->build($query);
+
+        self::assertSame(
+            'SELECT *' . ($expected === '' ? '' : ' WHERE ' . $this->replaceQuotes($expected)),
+            $sql,
+            'Generated SQL does not match expected SQL.',
+        );
+        self::assertSame(
+            $expectedParams,
+            $params,
+            'Bound parameters do not match expected parameters.',
+        );
+    }
+
+    public function testThrowInvalidArgumentExceptionWhenFromArrayDefinitionHasMissingOperands(): void
+    {
+        $this->expectException(
+            InvalidArgumentException::class,
+        );
+        $this->expectExceptionMessage(
+            "Operator 'IN' requires two operands.",
+        );
+
+        InCondition::fromArrayDefinition('IN', []);
+    }
+}